+++ go.mod
@@ -8,8 +8,10 @@ require (
 	github.com/aws/aws-lambda-go v1.26.0 // indirect
 	github.com/aws/aws-sdk-go v1.40.22 // indirect
 	github.com/caarlos0/env v3.5.0+incompatible
+	github.com/davyzhang/agw v0.0.0-20200908000401-cd09bf93af20 // indirect
 	github.com/go-sql-driver/mysql v1.5.0 // indirect
 	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/json-iterator/go v1.1.11 // indirect
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.0
 	go.mongodb.org/mongo-driver v1.7.2 // indirect
