+++ src/apitest/tags_test.go
@@ -1,9 +1,8 @@
 package apitest
 
 import (
-	"errors"
+	"encoding/json"
 	"fmt"
-	"net/http"
 	"testing"
 
 	"github.com/Sovianum/figma-search-app/src/client/clienttag"
@@ -21,9 +20,7 @@ func TestTagsTestSuite(t *testing.T) {
 }
 
 func (s *TagsTestSuite) TestCreateTags() {
-	cl := s.NewClient()
-
-	createTags(cl, fileid.New(), &clienttag.Tag{
+	s.createTags(fileid.New(), &clienttag.Tag{
 		ID:   tagid.New(),
 		Text: "tag1",
 	})
@@ -33,15 +30,13 @@ type tagsCreationRequest struct {
 	Tags []*clienttag.Tag `json:"tags"`
 }
 
-func createTags(c *Client, fileID fileid.ID, tags ...*clienttag.Tag) {
-	resp := NewQueryBuilder(c).
-		Pathf("/files/%s/tags/create", fileID).
-		WithRequestBody(tagsCreationRequest{
-			Tags: tags,
-		}).
-		Post()
+func (s *TagsTestSuite) createTags(fileID fileid.ID, tags ...*clienttag.Tag) {
+	b, err := json.Marshal(tagsCreationRequest{
+		Tags: tags,
+	})
+	s.Require().NoError(err)
 
-	if resp.Status != http.StatusOK {
-		panic(errors.New(fmt.Sprintf("unexpected response code %d", resp.Status)))
-	}
+	result, err := s.handler(s.Ctx, b)
+	s.Require().NoError(err)
+	fmt.Println(result)
 }
